requirements.txt
psycopg2==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0 